+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 - 06/02/21
 - Minor console output cleanup
 - Changed simulation to use commas in .csv output instead of tabs
+- Wrote Python script to analyse .csv output files + produces metrics (e.g. avg energies consumed per move)
 
 ## 0.1.16 - 04/02/21
 - Move settings of simulation to config file + change simulation to use that instead of passed-in arguments